+++ previews/impulse/popover_component_preview.rb
@@ -15,5 +15,10 @@ module Impulse
     def custom_header
       render_with_template
     end
+
+    # @display center true
+    def nested
+      render_with_template
+    end
   end
 end